+++ NEWS.md
@@ -1,3 +1,5 @@
+# cli (development version)
+
 
 # cli 2.4.0